Added Version 1.2 (Scripts) (Patch_001) 8-21-18

+++ scripts/entity_system/systems/ai/ai_melee_line_of_sight_system.lua
@@ -21,6 +21,7 @@ AIMeleeLineOfSightSystem.init = function (self, context, system_name)
 	self._world = context.world
 	self._physics_world = World.physics_world(self._world)
 	self._extensions = {}
+	self._frozen_extensions = {}
 	self._update_queue = {}
 	self._update_queue_index = 0
 	self._update_queue_length = 0
@@ -35,24 +36,44 @@ AIMeleeLineOfSightSystem.on_add_extension = function (self, world, unit, extensi
 
 	ScriptUnit.add_extension(nil, unit, extension_name, self.NAME)
 
-	local ext = ScriptUnit.extension(unit, self.NAME)
-	ext.index = index
-	self._update_queue[index] = ext
-	self._extensions[unit] = ext
+	local extension = ScriptUnit.extension(unit, self.NAME)
+	extension.index = index
+	self._update_queue[index] = extension
+	self._extensions[unit] = extension
 	self._update_queue_length = index
 
-	return ext
-end
-
-AIMeleeLineOfSightSystem.on_freeze_extension = function (self, unit, extension_name)
-	self:on_remove_extension(unit, extension_name)
+	return extension
 end
 
 AIMeleeLineOfSightSystem.on_remove_extension = function (self, unit, extension_name)
+	self._frozen_extensions[unit] = nil
+
+	self:_cleanup_extension(unit, extension_name)
 	ScriptUnit.remove_extension(unit, self.NAME)
+end
+
+AIMeleeLineOfSightSystem.on_freeze_extension = function (self, unit, extension_name)
+	local extension = self._extensions[unit]
+
+	fassert(extension, "Unit was already frozen.")
+
+	if extension == nil then
+		return
+	end
+
+	self._frozen_extensions[unit] = extension
+
+	self:_cleanup_extension(unit, extension_name)
+end
+
+AIMeleeLineOfSightSystem._cleanup_extension = function (self, unit, extension_name)
+	local extensions = self._extensions
+
+	if extensions[unit] == nil then
+		return
+	end
 
 	local queue = self._update_queue
-	local extensions = self._extensions
 	local last_index = self._update_queue_length
 	local last_item = queue[last_index]
 	local remove_index = extensions[unit].index
@@ -61,6 +82,32 @@ AIMeleeLineOfSightSystem.on_remove_extension = function (self, unit, extension_n
 	last_item.index = remove_index
 	extensions[unit] = nil
 	self._update_queue_length = self._update_queue_length - 1
+end
+
+AIMeleeLineOfSightSystem.freeze = function (self, unit, extension_name, reason)
+	local frozen_extensions = self._frozen_extensions
+
+	if self._frozen_extensions[unit] then
+		return
+	end
+
+	local extension = self._extensions[unit]
+
+	fassert(extension, "Unit to freeze didn't have unfrozen extension")
+	self:_cleanup_extension(unit, extension_name)
+
+	self._extensions[unit] = nil
+	frozen_extensions[unit] = extension
+end
+
+AIMeleeLineOfSightSystem.unfreeze = function (self, unit)
+	local extension = self._frozen_extensions[unit]
+	self._frozen_extensions[unit] = nil
+	local index = self._update_queue_length + 1
+	extension.index = index
+	self._update_queue[index] = extension
+	self._extensions[unit] = extension
+	self._update_queue_length = index
 end
 
 AIMeleeLineOfSightSystem.hot_join_sync = function (self, sender, player)
@@ -109,7 +156,7 @@ AIMeleeLineOfSightSystem.update = function (self, context, t)
 	while index <= max_index and raycasts < MAX_RAYCASTS do
 		local ext = queue[index]
 		local bb = ext.blackboard
-		local target = bb.attacking_target or bb.special_attacking_target or bb.target_unit
+		local target = bb.attacking_target or bb.target_unit
 
 		if unit_alive(target) and is_character(target) then
 			local unit = ext.unit
@@ -124,10 +171,6 @@ AIMeleeLineOfSightSystem.update = function (self, context, t)
 					local target_offset = target_pos - self_pos
 					success = false
 
-					if debug then
-						QuickDrawer:vector(self_pos, target_offset, Color(0, 255, 0))
-					end
-
 					if math.abs(target_offset.x) < BENEATH_ABOVE_EPSILON and math.abs(target_offset.y) < BENEATH_ABOVE_EPSILON then
 						raycasts = raycasts + 1
 						local z_offset_target = offsets[1].z
@@ -139,26 +182,14 @@ AIMeleeLineOfSightSystem.update = function (self, context, t)
 						if RAY_DISTANCE_EPSILON < dist then
 							local hit, hit_pos, hit_dist, hit_normal, hit_actor = PhysicsWorld.raycast(physics_world, from, dir, dist, "closest", "collision_filter", "filter_ai_line_of_sight_check")
 
-							if debug then
-								QuickDrawer:line(from, from + dist * dir, Color(0, 255, 255))
-							end
-
 							if not hit or Actor.unit(hit_actor) == target then
 								success = true
 							end
 						else
-							if debug then
-								QuickDrawer:line(from, from + 0.1 * dir, Color(255, 255, 255))
-							end
-
 							success = true
 						end
 					else
 						local right_vector = Vector3.normalize(Vector3.cross(target_offset, up))
-
-						if debug then
-							QuickDrawer:vector(self_pos, right_vector, Color(255, 0, 0))
-						end
 
 						for i = 1, num_offsets, 1 do
 							raycasts = raycasts + 1
@@ -177,10 +208,6 @@ AIMeleeLineOfSightSystem.update = function (self, context, t)
 								end
 
 								break
-							end
-
-							if debug then
-								QuickDrawer:line(from, from + dist * dir, Color(255, 0, 255))
 							end
 						end
 					end
